Fix useState drawer close issue and move historical sub functions to historical_storms_utils
