+++ react/components/header_nav.js
@@ -1,7 +1,7 @@
 import React, { useState } from "react";
 import { Box, IconButton, Drawer, List, ListItem, ListItemText } from "@mui/material";
 import MenuIcon from "@mui/icons-material/Menu";
-import { StormSearchQuery } from "./search_storm";
+import { StormSearchQuery } from "./search_storm_header";
 
 export default function HeaderNav({ children, navItems }) {
     const [headerDrawerOpen, setHeaderDrawerOpen] = useState(false);
@@ -87,9 +87,7 @@ export default function HeaderNav({ children, navItems }) {
                         </ListItem>
                     ))}
                 </List>
-                <Box className='search-drawer-small-screens'>
-                    <StormSearchQuery/>
-                </Box>
+                
                 
 
             </Drawer>
