+++ react/components/map.js
@@ -18,7 +18,7 @@ import StormPointDetails, { empty_point_obj } from "@/components/storm_point_det
 import { useDatasetDescriptions } from "@/pages/api/all_erddap_dataset";
 import { empty_station_obj } from "./layout";
 import StationDashboard from "./station_dashboard/station_dashboard";
-import { SearchSmallScreens } from "./search_small_screens";
+import { RenderStormSearch } from "./render_storm_search";
 //import StationDashboardTest from "./station_dashboard/station_dashboard";
 
 const defaultPosition = [46.9736, -54.69528]; // Mouth of Placentia Bay
@@ -55,7 +55,7 @@ export default function Map({ children, storm_points, storm_data, station_data, 
         )}
 
         {
-          <SearchSmallScreens
+          <RenderStormSearch
             isSearchSubmitted = {isSearchSubmitted}
             setIsSearchSubmitted= {setIsSearchSubmitted}
             searchResult= {searchResult}
