+++ react/components/search_storm_header.js
@@ -3,7 +3,7 @@ import { IconButton, TextField, Box, Typography, Paper } from "@mui/material";
 import Search from "@mui/icons-material/Search";
 import { useEffect, useState } from 'react';
 import { Stack } from "@mui/system";
-import {handleClick} from '../pages/about_page'
+import {handleStormNameClick} from '../pages/about_page'
 import { useRouter } from 'next/router';
 
 
@@ -93,7 +93,7 @@ export function RenderSearchResult({searchResult, router}){
           <Paper
             key={storm.storm_id}
             onClick={(e) => { console.log()
-                              handleClick(storm.name, storm.year, router);
+                              handleStormNameClick(storm.name, storm.year, router);
                               //triggerReload(); // Reload page when a storm is clicked
             
                               //console.log(storm);
